Add Panic Handling to ListUsers
