+++ pkg/server/commands/listusers/list_users_rpc_test.go
@@ -3,11 +3,13 @@ package listusers
 import (
 	"context"
 	"fmt"
+	"sync/atomic"
 	"testing"
 	"time"
 
 	"github.com/oklog/ulid/v2"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/goleak"
 	"go.uber.org/mock/gomock"
 
@@ -18,8 +20,11 @@ import (
 	"github.com/openfga/openfga/internal/mocks"
 	"github.com/openfga/openfga/internal/throttler/threshold"
 	"github.com/openfga/openfga/pkg/dispatch"
+	"github.com/openfga/openfga/pkg/logger"
+	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	"github.com/openfga/openfga/pkg/storage"
 	"github.com/openfga/openfga/pkg/storage/memory"
+	"github.com/openfga/openfga/pkg/storage/storagewrappers"
 	storagetest "github.com/openfga/openfga/pkg/storage/test"
 	"github.com/openfga/openfga/pkg/testutils"
 	"github.com/openfga/openfga/pkg/tuple"
@@ -4027,4 +4032,128 @@ func TestListUsersRespectsConsistency(t *testing.T) {
 		})
 		require.NoError(t, err)
 	})
+}
+
+func TestListUsersDirectRelationshipPanicListUsers(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+	model := `
+		model
+			schema 1.1
+		type user
+		type document
+			relations
+				define viewer: [user]`
+
+	tests := ListUsersTests{
+		{
+			name: "direct_relationship_panic_listusers_max_results",
+			req: &openfgav1.ListUsersRequest{
+				Object:   &openfgav1.Object{Type: "document", Id: "1"},
+				Relation: "viewer",
+				UserFilters: []*openfgav1.UserTypeFilter{
+					{
+						Type: "user",
+					},
+				},
+			},
+			model: model,
+			tuples: []*openfgav1.TupleKey{
+				tuple.NewTupleKey("document:1", "viewer", "user:will"),
+				tuple.NewTupleKey("document:1", "viewer", "user:maria"),
+				tuple.NewTupleKey("document:2", "viewer", "user:jon"),
+			},
+			expectedUsers: []string{},
+			//expectedErrorMsg:  ErrPanic.Error(),
+			newListUsersQuery: NewListUsersQueryPanicListUsersMaxResults,
+		},
+	}
+	tests.runListUsersTestCases(t)
+}
+
+func TestListUsersComputedRelationshipPanicListUsersExpand(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+	tests := ListUsersTests{
+		{
+			name: "computed_relationship_with_contextual_tuples_panic_listusers_expand",
+			req: &openfgav1.ListUsersRequest{
+				Object:   &openfgav1.Object{Type: "document", Id: "1"},
+				Relation: "viewer",
+				UserFilters: []*openfgav1.UserTypeFilter{
+					{
+						Type: "user",
+					},
+				},
+				ContextualTuples: []*openfgav1.TupleKey{
+					tuple.NewTupleKey("document:1", "owner", "user:will"),
+					tuple.NewTupleKey("document:1", "owner", "user:maria"),
+					tuple.NewTupleKey("document:2", "viewer", "user:jon"),
+				},
+			},
+			model: `
+				model
+					schema 1.1
+				type user
+				type document
+					relations
+						define owner: [user]
+						define viewer: owner`,
+			tuples:            []*openfgav1.TupleKey{},
+			expectedUsers:     []string{},
+			expectedErrorMsg:  ErrPanic.Error(),
+			newListUsersQuery: NewListUsersQueryPanicListUsersExpand,
+		},
+	}
+	tests.runListUsersTestCases(t)
+}
+
+func NewListUsersQueryPanicListUsersMaxResults(ds storage.RelationshipTupleReader, contextualTuples []*openfgav1.TupleKey, opts ...ListUsersQueryOption) *listUsersQuery {
+	l := &listUsersQuery{
+		logger:                  logger.NewNoopLogger(),
+		resolveNodeBreadthLimit: serverconfig.DefaultResolveNodeBreadthLimit,
+		resolveNodeLimit:        serverconfig.DefaultResolveNodeLimit,
+		deadline:                serverconfig.DefaultListUsersDeadline,
+		maxResults:              serverconfig.DefaultListUsersMaxResults,
+		maxConcurrentReads:      serverconfig.DefaultMaxConcurrentReadsForListUsers,
+		wasThrottled:            new(atomic.Bool),
+		listUsersMaxResults: func(foundUsersCh chan foundUser, foundUsersUnique map[tuple.UserString]foundUser, maxResults uint32, span trace.Span, doneWithFoundUsersCh chan struct{}) {
+			panic(ErrPanic)
+		},
+		listUsersExpand: listUsersExpand,
+	}
+
+	for _, opt := range opts {
+		opt(l)
+	}
+
+	l.datastore = storagewrappers.NewRequestStorageWrapper(ds, contextualTuples, l.maxConcurrentReads)
+
+	return l
+}
+
+func NewListUsersQueryPanicListUsersExpand(ds storage.RelationshipTupleReader, contextualTuples []*openfgav1.TupleKey, opts ...ListUsersQueryOption) *listUsersQuery {
+	l := &listUsersQuery{
+		logger:                  logger.NewNoopLogger(),
+		resolveNodeBreadthLimit: serverconfig.DefaultResolveNodeBreadthLimit,
+		resolveNodeLimit:        serverconfig.DefaultResolveNodeLimit,
+		deadline:                serverconfig.DefaultListUsersDeadline,
+		maxResults:              serverconfig.DefaultListUsersMaxResults,
+		maxConcurrentReads:      serverconfig.DefaultMaxConcurrentReadsForListUsers,
+		wasThrottled:            new(atomic.Bool),
+		listUsersMaxResults:     listUsersMaxResults,
+		listUsersExpand: func(cancellableCtx context.Context, l *listUsersQuery, req *openfgav1.ListUsersRequest, dispatchCount *atomic.Uint32, foundUsersCh chan foundUser) expandResponse {
+			panic(ErrPanic)
+		},
+	}
+
+	for _, opt := range opts {
+		opt(l)
+	}
+
+	l.datastore = storagewrappers.NewRequestStorageWrapper(ds, contextualTuples, l.maxConcurrentReads)
+
+	return l
 }
